types interface in good files
