--- src/lib/agent/plugins/erc20/interface/interface.ts
@@ -1,8 +1,0 @@
-import { Uint256 } from 'starknet';
-
-export interface ParamsValidationResult {
-    formattedSymbol: string;
-    formattedAddress: string;
-    tokenAddress: string;
-    formattedAmountUint256: Uint256;
-}
