+++ src/lib/agent/plugins/erc20/types/types.ts
@@ -1,3 +1,5 @@
+import { Uint256 } from 'starknet';
+
 /**
  * Formats token amount with correct decimals
  * @param {string} amount - Amount to format
@@ -10,3 +12,28 @@ export const DECIMALS = {
     DEFAULT: 18,
   };
   
+/**
+ * Result interface for transfer operations
+ * @interface TransferResult
+ */
+export interface TransferResult {
+  status: 'success' | 'failure';
+  amount?: string;
+  symbol?: string;
+  recipients_address?: string;
+  transaction_hash?: string;
+  error?: string;
+  step?: string;
+}
+
+
+/**
+ * Parameters validation result
+ * @interface ParamsValidationResult
+ */
+export interface ParamsValidationResult {
+    formattedSymbol: string;
+    formattedAddress: string;
+    tokenAddress: string;
+    formattedAmountUint256: Uint256;
+}
